skip pid teardown if pid_dir not set

+++ vit/pid_manager.py
@@ -17,15 +17,16 @@ class PidManager(object):
             self._write_pid_file()
 
     def teardown(self):
-        try:
-            os.remove(self.pid_file)
-        # TODO: This needs a little more work to skip errors when no PID file
-        #       exists.
-        #except OSError as e:
-        #    if e.errno != errno.ENOENT:
-        #        raise OSError("could not remove pid file %s" % self.pid_file)
-        except:
-            pass
+        if self.pid_dir:
+            try:
+                os.remove(self.pid_file)
+            # TODO: This needs a little more work to skip errors when no PID file
+            #       exists.
+            #except OSError as e:
+            #    if e.errno != errno.ENOENT:
+            #        raise OSError("could not remove pid file %s" % self.pid_file)
+            except:
+                pass
 
     def _format_pid_dir(self):
         config_pid_dir = self.config.get('vit', 'pid_dir')
